+++ src/shuriken/doo_test.cljs
@@ -1,8 +1,8 @@
 (ns shuriken.doo-test
   (:require [cljs.test :as test]
             [doo.runner :refer-macros [doo-tests]]
-            [shuriken.associative-test]))
+            #_[shuriken.associative-test]))
 
-(enable-console-print!)
-(doo-tests 'shuriken.associative-test
-           'shuriken.string-test)
+; (enable-console-print!)
+; (doo-tests 'shuriken.associative-test
+;            'shuriken.string-test)