+++ test/shuriken/byte_buddy_test.clj
@@ -6,9 +6,9 @@
 
 (copy-class! Mother 'XMother)
 (import 'XMother
+        'XMother$Infant
         'XMother$Child
-        'XMother$Crying
-        'XMother$Infant)
+        'XMother$Crying)
 
 ;; We need to test the case where there is no child classes.
 (copy-class! MonkeyPatched 'MonkeyPatchedTestByteBuddyCopy)
@@ -16,7 +16,8 @@
 (deftest test-copy-class!
   (let [mom (new XMother)
         c   (.giveBirth mom)]
-    (is (instance? XMother$Child  c))
+    (is (= XMother$Child (class c)))
+    (is (instance? XMother$Child c))
     (is (= [XMother$Crying]                        (interfaces XMother$Child)))
     (is (= #{Object XMother$Crying XMother$Infant} (ancestors XMother$Child))))