+++ test/shuriken/reader_macros/diaeresis_unquote_test.clj
@@ -12,12 +12,10 @@
 (defmacro test-diaeresis-unquote []
   `(do ¨the-object))
 
-(deftest test-diaeresis-unquote
+#_(deftest test-diaeresis-unquote
   (testing "asserting default behavior of ~"
     (is (true? (thrown? #"Can't embed object in code"
                         (with-ns 'shuriken.reader-macros.diaeresis-unquote-test
                           (test-unquote))))))
   (testing "overcoming ~ limitation "
     (is (= the-object (test-diaeresis-unquote)))))
-
-(run-tests)